requirements.txt
asgiref==3.3.1
Django==3.1.3
django-rest-utils==0.0.5
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
